# Backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Test API with swagger",
    "version": "1.0.0",
    "description": ""
  },
  "basePath": "/api",
...
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/user/get/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/assignment/get/subject/{id}": {
      "get": {
        "tags": [
          "Assignment"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "해당 과목의 퀴즈들을 성공적으로 가져올 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "assignments": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 0
                      },
                      "subject": {
                        "$ref": "#/definitions/subject"
                      },
                      "title": {
                        "type": "string",
                        "example": "인공지능 과제 #2"
                      },
                      "content": {
                        "type": "string",
                        "example": "ㅋㅋ"
                      },
                      "score": {
                        "type": "number",
                        "example": 10
                      },
                      "date": {
                        "type": "string",
                        "example": "2021-05-05T15:38:19.424Z"
                      },
                      "deadline": {
                        "type": "string",
                        "example": "2021-05-08T00:00:00.000Z"
                      },
                      "comments": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "$ref": "#/definitions/user"
                            },
                            "content": {
                              "type": "string",
                              "example": "진심 몇 일째냐고~~~~~"
                            },
                            "date": {
                              "type": "string",
                              "example": "2021-05-05T15:40:19.424Z"
                            }
                          }
                        }
                      },
                      "emotions": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "type": "number",
                              "example": 0
                            },
                            "emotion": {
                              "type": "string",
                              "example": "angry"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/assignment/get/{id}": {
      "get": {
        "tags": [
          "Assignment"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "해당 과제를 성공적으로 받아올 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "assignment": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "number",
                      "example": 0
                    },
                    "subject": {
                      "$ref": "#/definitions/subject"
                    },
                    "title": {
                      "type": "string",
                      "example": "인지과제 #2"
                    },
                    "content": {
                      "type": "string",
                      "example": "ㅋㅋ"
                    },
                    "date": {
                      "type": "string",
                      "example": "2021-05-05T15:38:19.424Z"
                    },
                    "deadline": {
                      "type": "string",
                      "example": "2021-05-08T00:00:00.000Z"
                    },
                    "comments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "user": {
                            "$ref": "#/definitions/user"
                          },
                          "content": {
                            "type": "string",
                            "example": "ㅋㅋ"
                          },
                          "date": {
                            "type": "string",
                            "example": "2021-05-05T15:40:19.424Z"
                          }
                        }
                      }
                    },
                    "emotions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "user": {
                            "type": "number",
                            "example": 1
                          },
                          "emotion": {
                            "type": "string",
                            "example": "smile"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "404": {
            "description": "해당하는 강의노트가 존재하지 않을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "existAssignment": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/assignment/create": {
      "post": {
        "tags": [
          "Assignment"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "subject": {
                  "type": "number",
                  "example": 0
                },
                "title": {
                  "type": "string",
                  "example": "오늘은 여기까지만..."
                },
                "content": {
                  "type": "string",
                  "example": "힘들드아"
                },
                "file": {
                  "type": "number",
                  "example": 0
                },
                "score": {
                  "type": "number",
                  "example": 10
                },
                "date": {
                  "type": "string",
                  "example": "2021-05-05"
                },
                "deadline": {
                  "type": "string",
                  "example": "2021-05-08"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "정상적으로 과제를 생성했을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "assignment": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "number",
                      "example": 0
                    },
                    "subject": {
                      "type": "number",
                      "example": 0
                    },
                    "title": {
                      "type": "string",
                      "example": "인지과제 #3"
                    },
                    "content": {
                      "type": "string",
                      "example": "ㅋㅅㅋ"
                    },
                    "file": {
                      "type": "number",
                      "example": 0
                    },
                    "score": {
                      "type": "number",
                      "example": 10
                    },
                    "date": {
                      "type": "string",
                      "example": "2021-05-05T15:38:19.424Z"
                    },
                    "deadline": {
                      "type": "string",
                      "example": "2021-05-08T00:00:00.000Z"
                    },
                    "comments": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    },
                    "emotions": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    },
                    "checked": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우",
            "schema": {
              "$ref": "#/definitions/proAuthFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/assignment/update": {
      "put": {
        "tags": [
          "Assignment"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 0
                },
                "title": {
                  "type": "string",
                  "example": "오늘은 여기까지만..."
                },
                "content": {
                  "type": "string",
                  "example": "힘들드아"
                },
                "file": {
                  "type": "number",
                  "example": 0
                },
                "score": {
                  "type": "number",
                  "example": 15
                },
                "deadline": {
                  "type": "string",
                  "example": "2021-05-08"
                },
                "date": {
                  "type": "string",
                  "example": "2021-05-05"
                },
                "checked": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "정상적으로 과제를 수정했을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "Assignment": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "number",
                      "example": 0
                    },
                    "subject": {
                      "type": "number",
                      "example": 0
                    },
                    "title": {
                      "type": "string",
                      "example": "오늘은 여기까지만..."
                    },
                    "content": {
                      "type": "string",
                      "example": "힘들드아"
                    },
                    "file": {
                      "type": "number",
                      "example": 0
                    },
                    "score": {
                      "type": "number",
                      "example": 15
                    },
                    "date": {
                      "type": "string",
                      "example": "2021-05-05T15:38:19.424Z"
                    },
                    "deadline": {
                      "type": "string",
                      "example": "2021-05-08T00:00:00.000Z"
                    },
                    "comments": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    },
                    "emotions": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    },
                    "checked": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우",
            "schema": {
              "$ref": "#/definitions/proAuthFailed"
            }
          },
          "404": {
            "description": "해당 Assignment가 존재하지 않는 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "existAssignment": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/assignment/delete/{id}": {
      "delete": {
        "tags": [
          "Assignment"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "정상적으로 과제를 삭제했을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우",
            "schema": {
              "$ref": "#/definitions/proAuthFailed"
            }
          },
          "404": {
            "description": "해당 Assignment가 존재하지 않는 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "existAssignment": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "authFailed": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "isLogined": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "proAuthFailed": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "isProfessor": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "subject": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "number",
          "example": 0
        },
        "name": {
          "type": "string",
          "example": "캡스톤디자인"
        },
        "start_period": {
          "type": "string",
          "example": "2021-03-02"
        },
        "end_period": {
          "type": "string",
          "example": "2021-06-30"
        },
        "start_time": {
          "type": "array",
          "example": [
            "16:30",
            "18:00",
            "19:30"
          ],
          "items": {
            "type": "string"
          }
        },
        "end_time": {
          "type": "array",
          "example": [
            "18:00",
            "19:30",
            "21:00"
          ],
          "items": {
            "type": "string"
          }
        },
        "days": {
          "type": "array",
          "example": [
            1,
            1,
            1
          ],
          "items": {
            "type": "number"
          }
        },
        "code": {
          "type": "string",
          "example": "519hi32hkjifb12"
        },
        "lectures": {
          "type": "array",
          "example": [],
          "items": {}
        },
        "students": {
          "type": "array",
          "example": [
            0
          ],
          "items": {
            "type": "number"
          }
        },
        "introURL": {
          "type": "string",
          "example": ""
        }
      }
    },
...
    "user": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "ㄱㄴㄷ"
        },
        "email": {
          "type": "string",
          "example": "amola@ajou.ac.kr"
        },
        "school": {
          "type": "string",
          "example": "아주대학교"
        },
        "identityID": {
          "type": "string",
          "example": "201620817"
        },
        "major": {
          "type": "string",
          "example": "컴퓨터공학"
        },
        "type": {
          "type": "string",
          "example": "student"
        },
        "subject": {
          "type": "array",
          "example": [
            0
          ],
          "items": {
            "type": "number"
          }
        }
      }
    }
  }
}